backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
gunicorn>=21.0.0
python-dotenv==1.0.0
pydantic
pydantic-settings==2.1.0
email-validator>=2.1.0
python-multipart==0.0.6

# Database - Supabase needs httpx >= 0.26
supabase
httpx>=0.26,<0.29

# Telegram Bot - Version 21+ supports httpx >= 0.26
python-telegram-bot>=21.0

# AI & ML - Python 3.12 Compatible
openai>=1.3.0
google-genai>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
joblib>=1.3.0

# HTTP & Requests
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.12.0
flake8>=6.1.0

# Logging
python-json-logger>=2.0.0

# Scheduling
schedule>=1.2.0

# Date handling
python-dateutil>=2.8.0

# Document processing
pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.13